batch files - upload into single directory, changed name of team photo storage location
Batch files that read the pit and match data files from the tablets now
put a copy of the files into single files for match and pit scouting in
addition to making tablet specific copies.
Changed the name of the subdirectory where team photos are stored
(python utility ConvertPitPhotos.py.